refactor: Enforce Code Style Rules on All Projects and Center Nuget Package Dependencies
Resolves: #12
