+++ .editorconfig
@@ -208,11 +208,10 @@ dotnet_analyzer_diagnostic.category-Style.severity = warning
 dotnet_analyzer_diagnostic.severity = warning
 
 # Specific diagnostic overrides
-dotnet_diagnostic.IDE0058.severity = none
+dotnet_diagnostic.IDE0058.severity = silent
 dotnet_diagnostic.CRRSP08.severity = none 
 dotnet_diagnostic.CA1000.severity = none
 dotnet_diagnostic.IDE0055.severity = none
-dotnet_diagnostic.IDE0058.severity = none
 dotnet_diagnostic.IDE2003.severity = none
 
 # Custom word list
