+++ src/Core/OnForkHub.Core/Validations/CategoryValidationService.cs
@@ -1,4 +1,5 @@
 using OnForkHub.Core.Interfaces.Validations;
+using OnForkHub.Core.Validations.Base;
 
 namespace OnForkHub.Core.Validations;
