+++ src/Shared/OnForkHub.Abstractions/OnForkHub.Abstractions.csproj
@@ -4,7 +4,6 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <GenerateResxSourceFilesAsPublic>true</GenerateResxSourceFilesAsPublic>
         <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
         <NoWarn>$(NoWarn);1591</NoWarn>
     </PropertyGroup>
