+++ src/Shared/OnForkHub.Dependencies/OnForkHub.Dependencies.csproj
@@ -9,6 +9,13 @@
     </PropertyGroup>
 
 
+
+    <!--Dependencies for ui projects-->
+    <ItemGroup>
+        <PackageReference Include="MudBlazor" Version="7.15.0" />
+    </ItemGroup>
+
+    <!--Dependencies for persitence projects-->
     <ItemGroup>
         <PackageReference Include="Dapper" Version="2.*" />
         <PackageReference Include="Microsoft.Data.SqlClient" Version="5.*" />
@@ -21,9 +28,19 @@
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.*" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.*" />
-        <PackageReference Include="MudBlazor" Version="7.15.0" />
+    </ItemGroup>
+
+
+    <!--Dependencies for test projects-->
+    <ItemGroup>
+        <PackageReference Include="FluentAssertions" Version="6.*" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="8.*" />
+        <PackageReference Include="NSubstitute" Version="5.*" />
+        <PackageReference Include="coverlet.collector" Version="6.*" />
+        <PackageReference Include="xunit" Version="2.*" />
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.*" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.*" />
     </ItemGroup>
 
 </Project>
