+++ test/Core/OnForkHub.Application.Test/OnForkHub.Application.Test.csproj
@@ -10,26 +10,14 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <Compile Remove="Services\NewFolder\**" />
-      <EmbeddedResource Remove="Services\NewFolder\**" />
-      <None Remove="Services\NewFolder\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="FluentAssertions" Version="6.*" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.10" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="8.0.10" />
-        <PackageReference Include="NSubstitute" Version="5.*" />
-        <PackageReference Include="xunit" Version="2.*" />
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.*" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.*" />
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\..\src\Core\OnForkHub.Application\OnForkHub.Application.csproj" />
-      <ProjectReference Include="..\..\..\src\Core\OnForkHub.Core\OnForkHub.Core.csproj" />
-      <ProjectReference Include="..\..\..\src\Shared\OnForkHub.Abstractions\OnForkHub.Abstractions.csproj" />
-      <ProjectReference Include="..\..\..\src\Shared\OnForkHub.Dependencies\OnForkHub.Dependencies.csproj" />
+        <ProjectReference Include="..\..\..\src\Core\OnForkHub.Application\OnForkHub.Application.csproj" />
+        <ProjectReference Include="..\..\..\src\Core\OnForkHub.Core\OnForkHub.Core.csproj" />
+        <ProjectReference Include="..\..\..\src\Shared\OnForkHub.Abstractions\OnForkHub.Abstractions.csproj" />
+        <ProjectReference Include="..\..\..\src\Shared\OnForkHub.Dependencies\OnForkHub.Dependencies.csproj" />
     </ItemGroup>
 
     <ItemGroup>
