+++ test/Core/OnForkHub.Application.Test/Validation/ValidationServiceTest.cs
@@ -1,3 +1,5 @@
+using OnForkHub.Core.Validations.Base;
+
 namespace OnForkHub.Application.Test.Validation;
 
 public class ValidationServiceTests
